Change TotalItems from int to long in GridData and TableData

+++ src/MudBlazor/Components/DataGrid/GridState.cs
@@ -78,6 +78,6 @@ namespace MudBlazor
         /// <remarks>
         /// This property is used to determine the number of pages of data.
         /// </remarks>
-        public int TotalItems { get; set; }
+        public long TotalItems { get; set; }
     }
 }
